IPython console: Move main widget enums to its API module
