+++ spyder/plugins/ipythonconsole/api.py
@@ -41,3 +41,57 @@ class IPythonConsolePyConfiguration(TypedDict):
     # If not None, then the console will use an alternative run method
     # (e.g. `runfile`, `debugfile` or `debugcell`).
     run_method: NotRequired[str]
+
+
+class IPythonConsoleWidgetActions:
+    # Clients creation
+    CreateNewClient = 'new tab'
+    CreateCythonClient = 'create cython client'
+    CreateSymPyClient = 'create cympy client'
+    CreatePyLabClient = 'create pylab client'
+    CreateNewClientEnvironment = 'create environment client'
+
+    # Current console actions
+    ClearConsole = 'Clear shell'
+    ClearLine = 'clear line'
+    ConnectToKernel = 'connect to kernel'
+    Interrupt = 'interrupt kernel'
+    InspectObject = 'Inspect current object'
+    Restart = 'Restart kernel'
+    ResetNamespace = 'reset namespace'
+    ShowEnvironmentVariables = 'Show environment variables'
+    ShowSystemPath = 'show system path'
+    ToggleElapsedTime = 'toggle elapsed time'
+    Quit = 'exit'
+
+    # Tabs
+    RenameTab = 'rename tab'
+
+    # Variables display
+    ArrayInline = 'enter array inline'
+    ArrayTable = 'enter array table'
+
+    # Documentation and help
+    IPythonDocumentation = 'ipython documentation'
+    ConsoleHelp = 'console help'
+    QuickReference = 'quick reference'
+
+
+class IPythonConsoleWidgetOptionsMenus:
+    SpecialConsoles = 'special_consoles_submenu'
+    Documentation = 'documentation_submenu'
+    EnvironmentConsoles = 'environment_consoles_submenu'
+
+
+class IPythonConsoleWidgetOptionsMenuSections:
+    Edit = 'edit_section'
+    View = 'view_section'
+
+
+class IPythonConsoleWidgetMenus:
+    TabsContextMenu = 'tabs_context_menu'
+
+
+class IPythonConsoleWidgetTabsContextMenuSections:
+    Consoles = 'tabs_consoles_section'
+    Edit = 'tabs_edit_section'
